threat-intel-agent/requirements.txt
# LangChain Deep Agents
deepagents>=0.1.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0

# HTTP & Feed Parsing
httpx>=0.27.0
feedparser>=6.0.0
aiofiles>=24.1.0

# Data Processing
pydantic>=2.0.0
python-dateutil>=2.9.0
jinja2>=3.1.0

# Optional: Web search tool
tavily-python>=0.5.0

# Optional: Email reports
# aiosmtplib>=3.0.0

# Scheduling (if not using systemd/cron)
apscheduler>=3.10.0

# Dev/Test
pytest>=8.0.0
pytest-asyncio>=0.24.0
python-dotenv>=1.0.0
